+++ src/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 using MelonLoader;
 using BoneAI;
 
-[assembly: MelonInfo(typeof(BoneAIMod), "BoneAI", "3.2.0", "BoneAI Community")]
+[assembly: MelonInfo(typeof(BoneAIMod), "BoneAI", "3.3.0", "BoneAI Community")]
 [assembly: MelonGame("Stress Level Zero", "BONELAB")]
 [assembly: MelonOptionalDependencies("LabFusion")]